+++ mayan/apps/document_states/tests/test_workflow_template_transition_field_api.py
@@ -1,3 +1,5 @@
+import json
+
 from rest_framework import status
 
 from mayan.apps.documents.tests.mixins.document_mixins import (
@@ -5,11 +7,19 @@ from mayan.apps.documents.tests.mixins.document_mixins import (
 )
 from mayan.apps.rest_api.tests.base import BaseAPITestCase
 
-from ..events import event_workflow_template_edited
+from ..events import (
+    event_workflow_instance_transitioned, event_workflow_template_edited
+)
 from ..permissions import (
+    permission_workflow_instance_transition,
     permission_workflow_template_edit, permission_workflow_template_view
 )
 
+from .literals import (
+    TEST_WORKFLOW_TEMPLATE_TRANSITION_FIELD_LOOKUP_CHOICE_DATA,
+    TEST_WORKFLOW_TEMPLATE_TRANSITION_FIELD_LOOKUP_CHOICE_LIST
+)
+from .mixins.workflow_instance_mixins import WorkflowInstanceAPIViewTestMixin
 from .mixins.workflow_template_transition_field_mixins import (
     WorkflowTemplateTransitionFieldAPIViewTestMixin
 )
@@ -227,6 +237,115 @@ class WorkflowTemplateTransitionFieldAPIViewTestCase(
         self.assertEqual(
             response.data['results'][0]['label'],
             self._test_workflow_template_transition_field.label
+        )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+
+class WorkflowTemplateTransitionFieldLookupAPIViewTestCase(
+    WorkflowInstanceAPIViewTestMixin,
+    WorkflowTemplateTransitionFieldAPIViewTestMixin, DocumentTestMixin,
+    BaseAPITestCase
+):
+    auto_upload_test_document = False
+
+    def setUp(self):
+        super().setUp()
+        self._create_test_workflow_template(add_test_document_type=True)
+        self._create_test_workflow_template_state()
+        self._create_test_workflow_template_state()
+        self._create_test_workflow_template_transition()
+
+    def test_workflow_instance_log_entries_create_api_view_with_extra_data_document_and_workflow_access(self):
+        self._create_test_workflow_template_transition_field(
+            extra_data={
+                'lookup': TEST_WORKFLOW_TEMPLATE_TRANSITION_FIELD_LOOKUP_CHOICE_DATA
+            }
+        )
+
+        self._create_test_document_stub()
+
+        self.grant_access(
+            obj=self._test_document,
+            permission=permission_workflow_instance_transition
+        )
+        self.grant_access(
+            obj=self._test_workflow_template,
+            permission=permission_workflow_instance_transition
+        )
+
+        self._clear_events()
+
+        TEST_WORKFLOW_INSTANCE_LOG_ENTRY_EXTRA_DATA_LOOKUP = json.dumps(
+            obj={
+                self._test_workflow_template_transition_field.name: TEST_WORKFLOW_TEMPLATE_TRANSITION_FIELD_LOOKUP_CHOICE_LIST[0]
+            }
+        )
+
+        response = self._request_test_workflow_instance_log_entry_create_api_view(
+            extra_data={
+                'extra_data': TEST_WORKFLOW_INSTANCE_LOG_ENTRY_EXTRA_DATA_LOOKUP
+
+            }, workflow_instance=self._test_workflow_instance
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        self.assertEqual(
+            self._test_workflow_instance.log_entries.count(), 1
+        )
+        self._test_workflow_instance.refresh_from_db()
+
+        self.assertEqual(
+            self._test_workflow_instance.get_context()['workflow_instance_context'],
+            json.loads(s=TEST_WORKFLOW_INSTANCE_LOG_ENTRY_EXTRA_DATA_LOOKUP)
+        )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+
+        self.assertEqual(events[0].action_object, self._test_document)
+        self.assertEqual(events[0].actor, self._test_case_user)
+        self.assertEqual(
+            events[0].target, self._test_document.workflows.first()
+        )
+        self.assertEqual(
+            events[0].verb, event_workflow_instance_transitioned.id
+        )
+
+    def test_workflow_instance_log_entries_create_api_bad_value_view_with_extra_data_document_and_workflow_access(self):
+        self._create_test_workflow_template_transition_field(
+            extra_data={
+                'lookup': TEST_WORKFLOW_TEMPLATE_TRANSITION_FIELD_LOOKUP_CHOICE_DATA
+            }
+        )
+
+        self._create_test_document_stub()
+
+        self.grant_access(
+            obj=self._test_document,
+            permission=permission_workflow_instance_transition
+        )
+        self.grant_access(
+            obj=self._test_workflow_template,
+            permission=permission_workflow_instance_transition
+        )
+
+        self._clear_events()
+
+        response = self._request_test_workflow_instance_log_entry_create_api_view(
+            extra_data={
+                'extra_data': json.dumps(
+                    obj={
+                        self._test_workflow_template_transition_field.name: 'bad-value'
+                    }
+                )
+            }, workflow_instance=self._test_workflow_instance
+        )
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+        self.assertEqual(
+            self._test_workflow_instance.log_entries.count(), 0
         )
 
         events = self._get_test_events()