+++ mayan/apps/document_states/views/workflow_instance_views.py
@@ -163,7 +163,8 @@ class WorkflowInstanceTransitionExecuteView(
                     'label': _(message='Comment'),
                     'class': 'django.forms.CharField', 'kwargs': {
                         'help_text': _(
-                            message='Optional comment to attach to the transition.'
+                            message='Optional comment to attach to the '
+                            'transition.'
                         ),
                         'required': False,
                     }
@@ -173,26 +174,67 @@ class WorkflowInstanceTransitionExecuteView(
                 'comment': {
                     'class': 'django.forms.widgets.Textarea',
                     'kwargs': {
-                        'attrs': {
-                            'rows': 3
-                        }
+                        'attrs': {'rows': 3}
                     }
                 }
             }
         }
 
-        for field in self.get_workflow_template_transition().fields.all():
+        workflow_template_transition = self.get_workflow_template_transition()
+
+        for field in workflow_template_transition.fields.all():
             schema['fields'][field.name] = {
                 'class': FIELD_TYPE_MAPPING[field.field_type],
                 'help_text': field.help_text,
                 'label': field.label,
                 'required': field.required
             }
+
+            schema_field = schema['fields'][field.name]
+
             if field.widget:
                 schema['widgets'][field.name] = {
                     'class': WIDGET_CLASS_MAPPING[field.widget],
                     'kwargs': field.get_widget_kwargs()
                 }
+
+            if field.lookup:
+                try:
+                    lookup_choices = field.get_lookup_values(
+                        workflow_instance=self.external_object
+                    )
+
+                    field_choices = list(
+                        zip(lookup_choices, lookup_choices)
+                    )
+
+                    if not field.required:
+                        field_choices.insert(
+                            0, ('', '------')
+                        )
+
+                    schema_field['class'] = 'django.forms.fields.ChoiceField'
+                    schema_field['kwargs'] = {'choices': field_choices}
+
+                    schema['widgets'][field.name] = {
+                        'class': 'django.forms.fields.Select',
+                        'kwargs': {
+                            'attrs': {'class': 'select2'}
+                        }
+                    }
+                except Exception as exception:
+                    schema_field['class'] = 'django.forms.fields.CharField'
+                    schema_field['kwargs'] = {
+                        'initial': _(
+                            message='Lookup value error: %s'
+                        ) % exception
+                    }
+                    schema['widgets'][field.name] = {
+                        'class': 'django.forms.fields.TextInput',
+                        'kwargs': {
+                            'attrs': {'readonly': 'readonly'}
+                        }
+                    }
 
         return {'schema': schema}